attached_assets/requirements_1749835961176.txt

fastapi==0.110.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
alembic==1.12.1
redis==5.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
pydantic==2.5.3
pydantic-settings==2.2.1
python-dotenv==1.0.0
httpx==0.27.0
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1
celery==5.3.4
flower==2.0.1
prometheus-client==0.19.0
structlog==23.2.0
rich==13.7.0
aiofiles==23.2.1
